+++ src/mesh/mesh_io.jl
@@ -89,12 +89,7 @@ function save_mesh_file(mesh::CurvedMesh, output_directory, timestep=0)
   # Create output directory (if it does not exist)
   mkpath(output_directory)
 
-  # Determine file name based on existence of meaningful time step
-  if timestep > 0
-    filename = joinpath(output_directory, @sprintf("mesh_%06d.h5", timestep))
-  else
-    filename = joinpath(output_directory, "mesh.h5")
-  end
+  filename = joinpath(output_directory, "mesh.h5")
 
   # Open file (clobber existing content)
   h5open(filename, "w") do file
@@ -102,8 +97,8 @@ function save_mesh_file(mesh::CurvedMesh, output_directory, timestep=0)
     attributes(file)["mesh_type"] = get_name(mesh)
     attributes(file)["ndims"] = ndims(mesh)
     attributes(file)["size"] = collect(size(mesh))
-    attributes(file)["coordinates_min"] = collect(mesh.coordinates_min)
-    attributes(file)["coordinates_max"] = collect(mesh.coordinates_max)
+
+    file["faces"] = mesh.faces .|> (f -> code_string(f, (Float64,))) .|> string
   end
 
   return filename
@@ -134,16 +129,14 @@ function load_mesh_serial(restart_file::AbstractString; n_cells_max)
     load_mesh!(mesh, restart_file)
   elseif mesh_type == "CurvedMesh"
     filename = get_restart_mesh_filename(restart_file, Val(false))
-    size_, coordinates_min_, coordinates_max_ = h5open(filename, "r") do file
+    size_, faces_string = h5open(filename, "r") do file
       return read(attributes(file)["size"]),
-             read(attributes(file)["coordinates_min"]),
-             read(attributes(file)["coordinates_max"])
+             read(file["faces"])
     end
 
     size = Tuple(size_)
-    coordinates_min = Tuple(coordinates_min_)
-    coordinates_max = Tuple(coordinates_max_)
-    mesh = CurvedMesh(size, coordinates_min, coordinates_max)
+    faces = faces_string .|> Meta.parse .|> eval
+    mesh = CurvedMesh(size, faces, Float64; unsaved_changes=false) # TODO RealT should be saved
   else
     error("Unknown mesh type!")
   end